adding reactive ideas to existing forest implementation
